Use CompletionItem.insertText by default, and fall back to label
This fixes cquery, where completions would be inserted with the function
signature included, which is extremely annoying.

It should however be noted that insertText is deprecated, and servers
should use textEdits instead, but thats a lot more involved in kakoune,
and luckily not a lot of servers seem to do that yet anyway.